requirements.txt
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
google-auth==1.23.0
idna==2.10
kubernetes==12.0.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0